requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Image processing
scikit-image>=0.21.0
scipy>=1.11.0

# AI/ML utilities
tqdm>=4.66.0
tensorboard>=2.14.0

# Configuration
PyYAML>=6.0

# CLI
argparse
colorama>=0.4.6

# Optional: GPU acceleration
# cupy-cuda11x>=12.0.0  # CUDA 11.x用
# cupy-cuda12x>=12.0.0  # CUDA 12.x用

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
